move: 태그 기능 패키지 이동 ( root/searchtag -> root/studyaddons/searchtag )
