--- server/edusync/src/main/java/com/codestates/edusync/searchtag/dto/SearchTagDto.java
@@ -1,19 +1,0 @@
-package com.codestates.edusync.searchtag.dto;
-
-import com.codestates.edusync.searchtag.entity.SearchTag;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import javax.validation.constraints.NotNull;
-import java.util.HashMap;
-import java.util.List;
-
-@NoArgsConstructor
-@Getter
-@Setter
-public class SearchTagDto {
-    private Long studygroupId;
-    @NotNull
-    private HashMap<String, String> tags;
-}
