+++ server/edusync/src/main/java/com/codestates/edusync/studyaddons/searchtag/dto/SearchTagDto.java
@@ -1,15 +1,17 @@
-package com.codestates.edusync.searchtag.dto;
+package com.codestates.edusync.studyaddons.searchtag.dto;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import javax.validation.constraints.NotNull;
 import java.util.HashMap;
-import java.util.List;
 
 @NoArgsConstructor
 @Getter
 @Setter
-public class SearchTagResponseDto {
+public class SearchTagDto {
+    private Long studygroupId;
+    @NotNull
     private HashMap<String, String> tags;
 }
