+++ server/edusync/src/main/java/com/codestates/edusync/studyaddons/searchtag/dto/SearchTagResponseDto.java
@@ -1,0 +1,14 @@
+package com.codestates.edusync.studyaddons.searchtag.dto;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import java.util.HashMap;
+
+@NoArgsConstructor
+@Getter
+@Setter
+public class SearchTagResponseDto {
+    private HashMap<String, String> tags;
+}
