+++ server/edusync/src/main/java/com/codestates/edusync/studyaddons/searchtag/entity/SearchTag.java
@@ -1,4 +1,4 @@
-package com.codestates.edusync.searchtag.entity;
+package com.codestates.edusync.studyaddons.searchtag.entity;
 
 import com.codestates.edusync.study.studygroup.entity.Studygroup;
 import lombok.Getter;
@@ -6,9 +6,6 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.*;
-
-import static javax.persistence.CascadeType.*;
-import static javax.persistence.FetchType.*;
 
 @NoArgsConstructor
 @Getter
